Addition commit

+++ brightnesschanger.cpp
@@ -9,8 +9,8 @@ QImage BrightnessChanger::change_brightness(const QImage& image) {
         return image;
     QImage new_image(image.size(), QImage::Format_RGB32);
     int64_t color;
-    for(uint64_t x = 0; x < image.width(); x++)
-        for(uint64_t y = 0; y < image.height(); y++) {
+    for(uint64_t x = 0; x < uint64_t(image.width()); x++)
+        for(uint64_t y = 0; y < uint64_t(image.height()); y++) {
             color = int64_t(image.pixelColor(x, y).red()) + _brightness_change;
             if(color < 0)
                 new_image.setPixel(x, y, qRgb(0, 0, 0));
