+++ contraster.cpp
@@ -10,8 +10,8 @@ QImage Contraster::change_contrast(const QImage &image) {
     QImage new_image = QImage(image.size(), QImage::Format_RGB32);
     uint64_t color;
     if(_decrease_mode) {
-        for(uint64_t x = 0; x < image.width(); x++) {
-            for(uint64_t y = 0; y < image.height(); y++) {
+        for(uint64_t x = 0; x < uint64_t(image.width()); x++) {
+            for(uint64_t y = 0; y < uint64_t(image.height()); y++) {
                 color = image.pixelColor(x, y).red();
                 if(color < _q1 || _q1 >= _q2)
                     color = 0;
@@ -24,8 +24,8 @@ QImage Contraster::change_contrast(const QImage &image) {
         }
     }
     else {
-        for(uint64_t x = 0; x < image.width(); x++) {
-            for(uint64_t y = 0; y < image.height(); y++) {
+        for(uint64_t x = 0; x < uint64_t(image.width()); x++) {
+            for(uint64_t y = 0; y < uint64_t(image.height()); y++) {
                 color = image.pixelColor(x, y).red();
                 if(color < _q1 || _q1 >= _q2)
                     color = 0;
