+++ negativetrashhold.cpp
@@ -7,8 +7,8 @@ NegativeTrashhold::NegativeTrashhold() {
 QImage NegativeTrashhold::negativize(const QImage &image) {
     QImage new_image = QImage(image.size(), QImage::Format_RGB32);
     uint64_t color;
-    for(uint64_t x = 0; x < image.width(); x++)
-        for(uint64_t y = 0; y < image.height(); y++) {
+    for(uint64_t x = 0; x < uint64_t(image.width()); x++)
+        for(uint64_t y = 0; y < uint64_t(image.height()); y++) {
             color = image.pixelColor(x, y).red();
             if(color < _trashhold)
                 new_image.setPixel(x, y, qRgb(color, color, color));
